+++ database/migrations/2025_04_02_092619_add_new_columns_to_release_release_format_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('release_release_format', function (Blueprint $table) {
-            $table->integer('price')->nullable();
+            $table->smallInteger('price')->nullable();
         });
     }
